Add Face Model to Hierarchy
Add the Glasses1 to Hierarchy. Fit the Glasses1 to the FaceModel. Make the Glasses1 as a child of the FaceModel.
Choose the face of the FaceModel and disable the Skinned Mesh Renderer.

Rename the FaceModel Object to Face Parent, and make it Prefab (when ask you what Prefab, select Original Prefab).