Create the new UI
Add at Canvas an Image as the Container to contains all the buttons.

Add to this Image the container image.

Add as a child of this Image a button for the "Buy Now" button. Add the Buy Now image/Sprite.

Duplicate Buy Now button and add the Cancel image/Sprite.

Add a new Image as a GlassesImage.

Add 2 buttons for the Colors of the glasses. Add the colors image/Sprite.

Add UIController Object and add the script.

Add to the UIController to the 3 variables to each one the Glasses 1 Info, Glasses 2 Info, Glasses 3 Info.

Add to the GlassesButtons the Enable functions, and to the GlassesInfo to the Cancel Button the Disable function. All of those are OnClick() button and add the UIController.