Update the UIController script
Add the methods to Enable and Disable the UI container with the info of the each one Glasses.

Add the methods to change the Material-Color of the each one Glasses. The material are into the Glasses Models > Tex & Mat file. For the others 2 are into the file Material.

Select the 2 buttons with colors into the Hierarchy of each one Glasses Info and add the OnClick() method to change the corol.

Add the URL to redirect when I click the Buy Now button for each one of the Glasses. Add to the buttons to OnClick() the GlassesBuy method.